typing module renamed to typechecking
