+++ aeon/__main__.py
@@ -10,7 +10,7 @@ from aeon.prelude.prelude import evaluation_vars
 from aeon.prelude.prelude import typing_vars
 from aeon.sugar.desugar import desugar
 from aeon.sugar.parser import parse_program
-from aeon.typing.typeinfer import check_type
+from aeon.typechecking.typeinfer import check_type
 from aeon.utils.ctx_helpers import build_context
 
 if __name__ == "__main__":
