+++ aeon/verification/horn.py
@@ -21,10 +21,10 @@ from aeon.core.types import t_bool
 from aeon.core.types import t_int
 from aeon.core.types import Top
 from aeon.core.types import Type
-from aeon.typing.context import EmptyContext
-from aeon.typing.context import TypingContext
-from aeon.typing.context import VariableBinder
-from aeon.typing.liquid import type_infer_liquid
+from aeon.typechecking.context import EmptyContext
+from aeon.typechecking.context import TypingContext
+from aeon.typechecking.context import VariableBinder
+from aeon.typechecking.liquid import type_infer_liquid
 from aeon.verification.helpers import constraint_builder
 from aeon.verification.helpers import end
 from aeon.verification.helpers import imp
